requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-genai
python-dotenv
Pillow
PyPDF2
pdf2image
requests

